API Passport Адмінка Покращення коду товарів, видалення медіа файлів Api збереження і доступу до медіа файлів Correct mistakes
Організація js css файлів
Налаштування webpack

+++ app/Services/Data/Media/DeleteMediaFileFromDb.php
@@ -11,6 +11,7 @@ class DeleteMediaFileFromDb
     public function handel($id)
     {
         $media = Media::findOrFail($id);
+        if ($media->products()->count() > 0) return false;
         $path = $media->path;
         $disc = $media->disc;
         $media->delete();
